requirements.txt
appdirs==1.4.4
asgiref==3.3.1
distlib==0.3.1
Django==3.1.3
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.0.4
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
virtualenv==20.1.0
